Implement support for handling OPTIONS message from client.
All DataStax drivers have added or are in the process of adding a
'heartbeat' feature that sends an 'OPTIONS' message on a predefined
interval to keep connections opened that may be timed out by ISPs or
intermediate networking equipment that times out idle connections.  If a
'SUPPORTED' message is not sent in response within the driver's
configured read timeout, the connection is closed and must be
re-established.

The java driver has implemented this feature in 2.0.10 and 2.1.5
(via JAVA-533) with a default idle timeout of 30 seconds.

Additionally, upgrades to driver version 2.0.10.