ProtocolVersion DataType codec support.
There are differences between what data types are supported at
difference protocol versions.  Newer versions support more types than
older versions naturally.  We also need to handle the case where a type
is removed or aliased, such as the case of the 'Text' type which was
aliased to the 'Varchar' type in version 3+.