Add integration tests for protocol v3 and v4 types
Also do a little bit of refactoring to facilitate common code driver21
and driver30 by creating a common-v3 module.  This was required because
TupleValue is part of the driver, so to prevent the common module from
requiring driver 2.1+ a separate module was needed.

Non-driver types like tinyint, smallint, date and time are accounted for
in common, with methods in driver20 for returning those returning
UnsupportedOperationException.